Factor tiled plugins into separate package.
