+++ .gitattributes
@@ -1,1 +1,2 @@
 databroker/_version.py export-subst
+.git_archival.txt  export-subst
