+++ bluesky-tiled-plugins/bluesky_tiled_plugins/_version.py
@@ -1,0 +1,17 @@
+# file generated by setuptools_scm
+# don't change, don't track in version control
+TYPE_CHECKING = False
+if TYPE_CHECKING:
+    from typing import Tuple, Union
+
+    VERSION_TUPLE = Tuple[Union[int, str], ...]
+else:
+    VERSION_TUPLE = object
+
+version: str
+__version__: str
+__version_tuple__: VERSION_TUPLE
+version_tuple: VERSION_TUPLE
+
+__version__ = version = "2.0.0b45.dev0+g43a742e7.d20240620"
+__version_tuple__ = version_tuple = (2, 0, 0, "dev0", "g43a742e7.d20240620")
