+++ requirements-client.txt
@@ -1,4 +1,4 @@
-mongoquery
+bluesky-tiled-plugins
 msgpack >=1.0.0
 orjson
 tiled[client] >=0.1.0-b13
