+++ setup.py
@@ -69,11 +69,6 @@ setup(
     python_requires='>={}'.format('.'.join(str(n) for n in min_version)),
     entry_points={
         "console_scripts": ["databroker = databroker.cli:main"],
-        "tiled.special_client": [
-            "CatalogOfBlueskyRuns = databroker.client:CatalogOfBlueskyRuns",
-            "BlueskyRun = databroker.client:BlueskyRun",
-            "BlueskyEventStream = databroker.client:BlueskyEventStream",
-        ],
     },
     classifiers=[
         'License :: OSI Approved :: BSD License',
